pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "negotiate-env"
version = "0.1.0"
description = "OpenEnv-compatible B2B SaaS contract negotiation environment"
readme = "negotiate_env/README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.1",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["negotiate_env*"]
